# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Roomly household management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /households/{householdId}/chores:
    get:
      operationId: listChores
      tags: [chores]
      summary: List chores for household
      parameters:
        - name: householdId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Chores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Chore"
    post:
      operationId: createChore
      tags: [chores]
      summary: Create a chore
      parameters:
        - name: householdId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChoreInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chore"
  /chores/{choreId}:
    patch:
      operationId: updateChore
      tags: [chores]
      summary: Update a chore
      parameters:
        - name: choreId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChoreUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chore"
    delete:
      operationId: deleteChore
      tags: [chores]
      summary: Delete a chore
      parameters:
        - name: choreId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /chores/{choreId}/complete:
    post:
      operationId: completeChore
      tags: [chores]
      summary: Mark a chore as completed
      parameters:
        - name: choreId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChoreCompletionInput"
      responses:
        "201":
          description: Completion recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChoreCompletion"
  /households/{householdId}/chores/schedule:
    get:
      operationId: getChoreSchedule
      tags: [chores]
      summary: Get the upcoming chore rotation schedule
      parameters:
        - name: householdId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Chore schedule
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChoreScheduleItem"

  # ── House Rules ───────────────────────────────────────────────────────────
...
  /households/{householdId}/dashboard:
    get:
      operationId: getHouseholdDashboard
      tags: [dashboard]
      summary: Get household dashboard summary
      parameters:
        - name: householdId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HouseholdDashboard"
  /households/{householdId}/expenses/summary:
    get:
      operationId: getExpenseSummary
      tags: [dashboard]
      summary: Get expense breakdown by category
      parameters:
        - name: householdId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Expense summary
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ExpenseCategoryTotal"
  /households/{householdId}/balances:
    get:
      operationId: getMemberBalances
      tags: [dashboard]
      summary: Get who owes what across all expenses
      parameters:
        - name: householdId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Member balances
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MemberBalance"

  # ── Landlord ──────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Expense:
      type: object
      required: [id, householdId, title, amount, category, paidByMemberId, splitType, settled, date, createdAt]
      properties:
        id:
          type: integer
        householdId:
          type: integer
        title:
          type: string
        amount:
          type: number
        category:
          type: string
          enum: [groceries, utilities, internet, streaming, household, transport, other]
        paidByMemberId:
          type: integer
        splitType:
          type: string
          enum: [equal, custom]
        settled:
          type: boolean
        notes:
          type: ["string", "null"]
        date:
          type: string
        createdAt:
          type: string
        splits:
          type: array
          items:
            $ref: "#/components/schemas/ExpenseSplit"
    ExpenseSplit:
      type: object
      required: [id, expenseId, memberId, amount, paid]
      properties:
        id:
          type: integer
        expenseId:
          type: integer
        memberId:
          type: integer
        amount:
          type: number
        paid:
          type: boolean
...
    Chore:
      type: object
      required: [id, householdId, title, frequency, rotationOrder, currentAssigneeMemberId, createdAt]
      properties:
        id:
          type: integer
        householdId:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        frequency:
          type: string
          enum: [daily, weekly, biweekly, monthly]
        rotationOrder:
          type: array
          items:
            type: integer
        currentAssigneeMemberId:
          type: integer
        nextDueDate:
          type: ["string", "null"]
        lastCompletedAt:
          type: ["string", "null"]
        createdAt:
          type: string
    ChoreInput:
      type: object
      required: [title, frequency, rotationOrder, currentAssigneeMemberId]
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        frequency:
          type: string
          enum: [daily, weekly, biweekly, monthly]
        rotationOrder:
          type: array
          items:
            type: integer
        currentAssigneeMemberId:
          type: integer
        nextDueDate:
          type: string
    ChoreUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        frequency:
          type: string
          enum: [daily, weekly, biweekly, monthly]
        rotationOrder:
          type: array
          items:
            type: integer
        currentAssigneeMemberId:
          type: integer
        nextDueDate:
          type: string
    ChoreCompletionInput:
      type: object
      required: [completedByMemberId]
      properties:
        completedByMemberId:
          type: integer
        notes:
          type: string
    ChoreCompletion:
      type: object
      required: [id, choreId, completedByMemberId, completedAt]
      properties:
        id:
          type: integer
        choreId:
          type: integer
        completedByMemberId:
          type: integer
        notes:
          type: ["string", "null"]
        completedAt:
          type: string
    ChoreScheduleItem:
      type: object
      required: [choreId, choreTitle, assigneeMemberId, assigneeName, dueDate, frequency]
      properties:
        choreId:
          type: integer
        choreTitle:
          type: string
        assigneeMemberId:
          type: integer
        assigneeName:
          type: string
        dueDate:
          type: string
        frequency:
          type: string
        overdue:
          type: boolean

    # ── House Rules ───────────────────────────────────────────────────────
...
    HouseholdDashboard:
      type: object
      required: [householdId, totalRent, memberCount, rentDueDate, currentMonthPayments, pendingExpenseTotal, overdueChores, upcomingChores, recentExpenses]
      properties:
        householdId:
          type: integer
        totalRent:
          type: number
        memberCount:
          type: integer
        rentDueDate:
          type: ["string", "null"]
        currentMonthPayments:
          type: object
          properties:
            paid:
              type: number
            pending:
              type: number
            late:
              type: number
          required: [paid, pending, late]
        pendingExpenseTotal:
          type: number
        overdueChores:
          type: integer
        upcomingChores:
          type: integer
        recentExpenses:
          type: array
          items:
            $ref: "#/components/schemas/Expense"
    ExpenseCategoryTotal:
      type: object
      required: [category, total, count]
      properties:
        category:
          type: string
        total:
          type: number
        count:
          type: integer
    MemberBalance:
      type: object
      required: [memberId, memberName, avatarColor, totalOwed, totalOwes, netBalance]
      properties:
        memberId:
          type: integer
        memberName:
          type: string
        avatarColor:
          type: ["string", "null"]
        totalOwed:
          type: number
        totalOwes:
          type: number
        netBalance:
          type: number

    # ── Landlord ──────────────────────────────────────────────────────────
...