# lib/api-zod/src/generated/types/expense.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Roomly household management API
 * OpenAPI spec version: 0.1.0
 */
import type { ExpenseCategory } from './expenseCategory';
import type { ExpenseSplit } from './expenseSplit';
import type { ExpenseSplitType } from './expenseSplitType';

export interface Expense {
  id: number;
  householdId: number;
  title: string;
  amount: number;
  category: ExpenseCategory;
  paidByMemberId: number;
  splitType: ExpenseSplitType;
  settled: boolean;
  /** @nullable */
  notes?: string | null;
  date: string;
  createdAt: string;
  splits?: ExpenseSplit[];
}